@@ -7,7 +7,7 @@ name = "source_modelling"
 authors = [{ name = "QuakeCoRE" }]
 description = "Source modelling library"
 readme = "README.md"
-requires-python = ">=3.10"
+requires-python = ">=3.11"
 dynamic = ["version"]
 dependencies = [
    "numpy",
