@@ -12,7 +12,7 @@ read_gsf(gsf_filepath)
 References
 ----------
 The GSF file format is used to define the grid of the source model for a fault.
-See https://wiki.canterbury.ac.nz/display/QuakeCore/File+Formats+Used+On+GM
+See https://ucdigitalsms.atlassian.net/wiki/spaces/QuakeCore/pages/3291694768/File+Formats+Used+In+Ground+Motion+Simulation#FileFormatsUsedInGroundMotionSimulation-SRFinfoformat
 for details on the GSF format.
 """
 
